updated custom event tree
